+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2155,7 +2155,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01M2N4H2J576TZH5CGJ409FE0A"
+    id: Optional[str] = "01M2NBZVYFV2RR2KC82W3A7R8R"
 
     description: Optional[str] = None